% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_to_event.R
\name{add_time_to_event}
\alias{add_time_to_event}
\title{Calculate time-to-event for aggressive sclerosis phenotype}
\usage{
add_time_to_event(d0)
}
\arguments{
\item{d0}{A tibble containing basic demographic variables.}

\item{d1}{A tibble containing processed EDSS data.}
}
\value{
A tibble
}
\description{
Based on data of patients who
}
\examples{
\dontrun{
p <- here::here("data-raw", "some-cool-file-name.xlsx")
d <- prepare_data(p)
outcome_data <- determine_aggressive_phenotype(
  d$id, d$relapses, d$edss, eye_check = TRUE
)
}
}
