% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ids.R
\name{extract_ids}
\alias{extract_ids}
\title{Extract patients with signed informed consent}
\usage{
extract_ids(path)
}
\arguments{
\item{path}{A character string specifying the path
to the Excel file containing the data.}
}
\value{
Integer vector with valid patients IDs
}
\description{
Reads patient IDs from an Excel file containing
a single column with valied ids.
}
\examples{
\dontrun{
p <- here::here("data-raw", "patients_with_informed_consent.xlsx")
incl_ids <- extract_ids(p)
}

}
