% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missingness.R
\name{plot_missingness}
\alias{plot_missingness}
\title{Visualising missing data}
\usage{
plot_missingness(d0)
}
\arguments{
\item{d0}{Tibble containing preprocessed data
as generated by \code{preprocess_predictors}.}
}
\value{
{ggplot2} figure with patients on
ordinate, predictors on abscissa, diveded
by the aggressive MS diagnosis.
}
\description{
Plots patient-by-predictor grid showing
patient/predictor pairs with vs. without
missing data.
}
