% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_predictors_mri.R
\name{preprocess_predictors_mri}
\alias{preprocess_predictors_mri}
\title{Preprocess MRI data.}
\usage{
preprocess_predictors_mri(d, onset)
}
\arguments{
\item{d}{A tibble with the original treatment data.}

\item{onset}{A tibble containing id/disease onset pairs.}
}
\value{
A tibble with data.
}
\description{
Preprocess MRI data.
}
