% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_predictors_relapses.R
\name{preprocess_predictors_relapses}
\alias{preprocess_predictors_relapses}
\title{Preprocess relapse counts.}
\usage{
preprocess_predictors_relapses(d, onset)
}
\arguments{
\item{d}{A tibble with the original relapse data.}

\item{onset}{A tibble containing id/disease onset pairs.}
}
\value{
A tibble with data.
}
\description{
Preprocess relapse counts.
}
