% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_predictors_treatment.R
\name{preprocess_predictors_treatment}
\alias{preprocess_predictors_treatment}
\title{Preprocess treatment variables.}
\usage{
preprocess_predictors_treatment(d, onset)
}
\arguments{
\item{d}{A tibble with the original treatment data.}

\item{onset}{A tibble containing id/disease onset pairs.}
}
\value{
A tibble with data.
}
\description{
Preprocess treatment variables.
}
