% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_typos.R
\name{recode_typos}
\alias{recode_typos}
\title{Manually recode typos in disease onset}
\usage{
recode_typos(data)
}
\arguments{
\item{data}{A tibble used for determining the
aggressive phenotype.}
}
\value{
A tibble identical to the input but with
corrected disease duration values and invalid
data removed.
}
\description{
Takes patient data and manually corrects
known typos in disease onset or duration
information. It is intended to be used
within the \code{determine_aggressive_phenotype}
workflow.
}
\details{
Additionally, the function drops invalid data
such as EDSS assessments recorded before the
documented disease onset.
}
