% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_data_structure.R
\name{specify_data_structure}
\alias{specify_data_structure}
\title{Specify structure for data import}
\usage{
specify_data_structure(show = TRUE)
}
\arguments{
\item{show}{Logical; if \code{TRUE} (default),
prints the list containing the expected data
structure. If \code{FALSE}, does not print it.}
}
\value{
A list:
\describe{
\item{sheet names}{Expected names of Excel sheets to import.}
\item{labels}{Labels to be used within the dataset.}
\item{col_types}{Specifications of expected column types.}
}
}
\description{
Defines the expected structure of the source
data file. Intended to be used inside the
\code{prepare_data} wrapper and should be
consulted whenever the structure of the
source data file changes.
}
