% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_predictors.R
\name{preprocess_predictors}
\alias{preprocess_predictors}
\title{Preprocess potential predictors.}
\usage{
preprocess_predictors(d0, t, r, m, c, o, chol)
}
\arguments{
\item{d0}{A tibble containing data pre-processed by
\code{determine_aggressive_phenotype}.}

\item{t}{A tibble containing treatment variables.}

\item{r}{A tibble containing relapses data.}

\item{m}{A tibble containing MRI data.}

\item{c}{A tibble containing CSF data.}

\item{o}{A tibble containing OCB data.}

\item{chol}{A tibble containing cholesterol data.}
}
\value{
A tibble with data.
}
\description{
Preprocess potential predictors.
}
\examples{
\dontrun{
raw_data <- here::here("data-raw", "some-cool-file-name.xlsx") |>
  prepare_data()
clas_data <- determine_aggressive_phenotype(
  d$id, d$relapses, d$edss, eye_check = TRUE
)
df <- preprocess_predictors(
  d0 = clas_data$data,
  t = raw_data$treatment,
  r = raw_data$relapses,
  m = raw_data$mri,
  c = raw_data$csf,
  o = raw_data$ocb,
  chol = raw_data$cholesterol
)
}

}
